+++ src/main/java/npc/entity/Personality.java
@@ -6,7 +6,7 @@ import lombok.Data;
 @Data
 @Builder
 public class Personality {
-private Long personalityPk;
+private Integer personalityPk;
 private PersonalityType personalityId;
 private String traits;
 private String description;
